moving code to api/application
