+++ Source/SelfService/Web/businessMoments/businessMomentEditor.tsx
@@ -11,7 +11,7 @@ import { Pivot, PivotItem, IDropdownOption, DefaultButton } from '@fluentui/reac
 
 import CodeEditor, { loader } from '@monaco-editor/react';
 import { saveBusinessmoment } from '../stores/businessmoment';
-import { HttpResponseApplication } from '../api/api';
+import { HttpResponseApplication } from '../api/application';
 import { BusinessMoment, HttpInputBusinessMoment, BusinessMomentEntity } from '../api/index';
