+++ Source/SelfService/Web/businessMoments/entityEditor.tsx
@@ -17,7 +17,7 @@ import { Text, Pivot, PivotItem, IDropdownOption } from '@fluentui/react';
 
 
 import CodeEditor, { loader } from '@monaco-editor/react';
-import { HttpResponseApplication } from '../api/api';
+import { HttpResponseApplication } from '../api/application';
 import { BusinessMomentEntity, HttpInputBusinessMomentEntity } from '../api/index';
 import { saveBusinessmomentEntity } from '../stores/businessmoment';
