+++ Source/SelfService/Web/insights/runtimeStatsV1/runtimeV1Stats.tsx
@@ -5,7 +5,7 @@ import React, { useEffect, useState } from 'react';
 import { Pivot, PivotItem, IDropdownOption } from '@fluentui/react';
 import { Dropdown } from '@fluentui/react/lib/Dropdown';
 
-import { HttpResponseApplication } from '../../api/api';
+import { HttpResponseApplication } from '../../api/application';
 import { getRuntimeV1 } from '../../api/insights';
 import { DumpJson } from '../dumpJson';
 import { FaillingPartitionsSummary } from './failingPartitionsSummary';
