+++ Source/SelfService/Web/microservice/microserviceViewScreen.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { HttpResponseApplication } from '../api/api';
+import { HttpResponseApplication } from '../api/application';
 import { Overview } from './view';
 
 type Props = {
