+++ Source/SelfService/Web/microservice/view.tsx
@@ -5,12 +5,13 @@ import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useReadable } from 'use-svelte-store';
 
-import { getPodStatus, HttpResponsePodStatus, HttpResponseApplication } from '../api/api';
+import { getPodStatus, HttpResponsePodStatus } from '../api/api';
 import { microservices } from '../stores/microservice';
 import { View as BaseView } from './base/view';
 import { View as BusinessMomentsAdaptorView } from './businessMomentsAdaptor/view';
 import { View as RawDataLogView } from './rawDataLog/view';
 import { View as PurchaseOrderApiView } from './purchaseOrder/view';
+import { HttpResponseApplication } from '../api/application';
 
 type Props = {
     application: HttpResponseApplication
