+++ Source/SelfService/Web/stores/microservice.ts
@@ -8,10 +8,9 @@ import {
     saveMicroservice as apiSaveMicroservice,
     MicroserviceInfo,
     getMicroservices,
-    getApplication,
     HttpResponseMicroservices,
-    HttpResponseApplication,
 } from '../api/api';
+
 import {
     MicroserviceSimple,
     MicroserviceBusinessMomentAdaptor,
@@ -20,6 +19,8 @@ import {
     MicroservicePurchaseOrder,
     MicroserviceIngressPath,
 } from '../api/index';
+
+import { getApplication, HttpResponseApplication } from '../api/application';
 
 export type MicroserviceStore = {
     edit: any;
